bootloader: reorganize the security helper functions
Move the new security function helper code into dedicated source
file (`pyi_security.c`). Split the original helper function into
separate helper for verifying the parent process executable
(`pyi_security_verify_parent_proces()`) and a separate helper for
verifying the ownership and permissions of the application's top-level
directory (`pyi_security_verify_application_home_dir`).

This will allow us to apply these security checks in multiple
places, with required granularity (i.e., in some cases, we can
verify only application's top-level directory).
